Publication Parser Function
A parsing function for the publication page. It uses BS4 to extract html elements by classes and IDs and extracts the required publication data from them and stores them in the publication format template. It specifically extracts the direct access link to the paper itself, which was missing in the author publications page.